+++ src/include/Graphics/EVE/Types.h
@@ -94,13 +94,13 @@ template <uint8_t precision> int operator/(int value, FixedTemplate<precision> d
  * @brief Details for loaded bitmap instance
  */
 struct BitmapSlot {
-	uint32_t address : 24; ///< Location of asset in ROM/RAMG
-	EVE::BitmapFormat format : 8;
-	uint32_t metrics : 24; ///< Location of loaded font metrics for soft fonts
-	uint8_t stride;
-	uint8_t width;
-	uint8_t height;
-	AssetID id; ///< Associated asset, 0 is 'unassigned'
+	uint32_t address : 24;		  ///< Location of asset in ROM/RAMG
+	EVE::BitmapFormat format : 8; // :5
+	uint32_t metrics : 24;		  ///< Location of loaded font metrics for soft fonts
+	uint16_t stride;			  // :12
+	uint16_t width;				  // :11
+	uint16_t height;			  // :11
+	AssetID id;					  ///< Associated asset, 0 is 'unassigned'
 };
 
 /**
